Refactorizacion, estado global anime y bugs
